use sx style instead of css
